# Make exact matcher look in full text