--- a/reach/airflow/dags/policy.py
+++ b/reach/airflow/dags/policy.py
@@ -83,7 +83,7 @@ def to_s3_model(*args):
     ) % '/'.join(args)
 
 
-def create_org_pipeline(dag, organisation, item_limits, spider_years):
+def create_org_pipeline(dag, organisation, item_limits, spider_years, epmc_metadata_key=None):
     """ Creates all tasks tied to a single organisation::
 
         Spider -> ParsePdf
@@ -140,13 +140,12 @@ def create_org_pipeline(dag, organisation, item_limits, spider_years):
 
     exactMatchRefs = exact_match_refs_operator.ExactMatchRefsOperator(
         task_id='ExactMatchRefs.%s' % organisation,
-        es_hosts=get_es_hosts(),
-        src_s3_key=parsePdf.dst_s3_key,
-        dst_s3_key=to_s3_output(
+        es_host=get_es_hosts(),
+        publications_path=epmc_metadata_key,
+        exact_matched_references_path=to_s3_output(
             dag, 'exact-matched-refs', organisation, '.json.gz'),
-        es_index=esIndexFullTexts,
-        dag=dag,
-        )
+        es_full_text_index=esIndexFullTexts,
+        dag=dag)
     
     esIndexFuzzyMatched = es_index_fuzzy_matched.ESIndexFuzzyMatchedCitations(
         task_id="ESIndexFuzzyMatchedCitations.%s" % organisation,
@@ -197,6 +196,7 @@ def create_dag(dag_id, default_args, spider_years,
             organisation,
             item_limits,
             spider_years,
+            epmc_metadata_key,
         )
         esIndexPublications >> fuzzyMatchRefs
         esIndexPublications >> exactMatchRefs