--- a/reach/airflow/dags/policy.py
+++ b/reach/airflow/dags/policy.py
@@ -83,7 +83,7 @@ def to_s3_model(*args):
     ) % '/'.join(args)
 
 
-def create_org_pipeline(dag, organisation, item_limits, spider_years, epmc_metadata_key=None):
+def create_org_pipeline(dag, organisation, item_limits, spider_years, publications_path=None):
     """ Creates all tasks tied to a single organisation::
 
         Spider -> ParsePdf
@@ -141,10 +141,10 @@ def create_org_pipeline(dag, organisation, item_limits, spider_years, epmc_metad
     exactMatchRefs = exact_match_refs_operator.ExactMatchRefsOperator(
         task_id='ExactMatchRefs.%s' % organisation,
         es_hosts=get_es_hosts(),
-        publications_path=epmc_metadata_key,
+        publications_path=publications_path,
         exact_matched_references_path=to_s3_output(
             dag, 'exact-matched-refs', organisation, '.json.gz'),
-        es_full_text_index=esIndexFullTexts,
+        es_full_text_index='-'.join([dag.dag_id, 'docs']),
         dag=dag)
 
     esIndexFuzzyMatched = es_index_fuzzy_matched.ESIndexFuzzyMatchedCitations(
@@ -182,6 +182,11 @@ def create_dag(dag_id, default_args, spider_years,
         'output', 'open-research', 'epmc-metadata', 'epmc-metadata.json.gz'
     ])
 
+    publications_path = '/'.join([
+        "datalabs-staging", "airflow",
+        "output", "open-research", "epmc-metadata", "epmc-metadata.json.gz"
+    ])
+
     esIndexPublications = es_index_epmc_metadata.ESIndexEPMCMetadata(
         task_id='ESIndexEPMCMetadata',
         src_s3_key=epmc_metadata_key,
@@ -196,7 +201,7 @@ def create_dag(dag_id, default_args, spider_years,
             organisation,
             item_limits,
             spider_years,
-            epmc_metadata_key,
+            publications_path,
         )
         esIndexPublications >> fuzzyMatchRefs
         esIndexPublications >> exactMatchRefs