--- a/reach/airflow/tasks/es_index_fulltext_docs.py
+++ b/reach/airflow/tasks/es_index_fulltext_docs.py
@@ -21,13 +21,13 @@ class ESIndexFulltextDocs(BaseOperator):
     )
 
     @apply_defaults
-    def __init__(self, src_s3_key, es_host, organisation, es_port=9200,
+    def __init__(self, src_s3_key, es_hosts, organisation, es_port=9200,
                  max_items=None, aws_conn_id='aws_default',
                  es_index=None, *args, **kwargs):
         """
         Args:
             src_s3_key: S3 URL for the json.gz output file.
-            es_host: the hostname of elasticsearch database.
+            es_hosts: the hostname of elasticsearch database.
             es_port: the port of elasticsearch database. Default to 9200.
             max_items: Maximum number of fulltexts to process.
             aws_conn_id: Aws connection name.
@@ -39,7 +39,7 @@ class ESIndexFulltextDocs(BaseOperator):
             raise ValueError('src_s3_key must end in .json.gz')
 
         self.src_s3_key = src_s3_key
-        self.es_host = es_host
+        self.es_hosts = es_hosts
         self.es_port = es_port
         self.organisation = organisation
         self.aws_conn_id = aws_conn_id
@@ -48,7 +48,7 @@ class ESIndexFulltextDocs(BaseOperator):
 
     def execute(self, context):
         es = reach.elastic.common.connect(
-            self.es_host, self.es_port)
+            self.es_hosts, self.es_port)
         s3 = WellcomeS3Hook()
 
         # TODO: implement skipping mechanism