--- a/reach/airflow/tasks/es_index_fulltext_docs.py
+++ b/reach/airflow/tasks/es_index_fulltext_docs.py
@@ -48,7 +48,7 @@ class ESIndexFulltextDocs(BaseOperator):
 
     def execute(self, context):
         es = reach.elastic.common.connect(
-            self.es_hosts, self.es_port)
+            self.es_hosts)
         s3 = WellcomeS3Hook()
 
         # TODO: implement skipping mechanism