--- a/reach/airflow/tasks/es_index_fuzzy_matched.py
+++ b/reach/airflow/tasks/es_index_fuzzy_matched.py
@@ -23,13 +23,13 @@ class ESIndexFuzzyMatchedCitations(BaseOperator):
     )
 
     @apply_defaults
-    def __init__(self, src_s3_key, es_host, organisation,  es_port=9200,
+    def __init__(self, src_s3_key, es_hosts, organisation,  es_port=9200,
                  item_limits=None, aws_conn_id='aws_default',
                  es_index=None, *args, **kwargs):
         """
         Args:
             src_s3_key: S3 URL for the json.gz output file.
-            es_host: the hostname of elasticsearch database.
+            es_hosts: the hostname of elasticsearch database.
             es_port: the port of elasticsearch database. Default to 9200.
             item_limits: Maximum number of citations to process.
             aws_conn_id: Aws connection name.
@@ -41,7 +41,7 @@ class ESIndexFuzzyMatchedCitations(BaseOperator):
             raise ValueError('src_s3_key must end in .json.gz')
 
         self.src_s3_key = src_s3_key
-        self.es_host = es_host
+        self.es_hosts = es_hosts
         self.es_port = es_port
         self.aws_conn_id = aws_conn_id
         self.item_limits = item_limits
@@ -50,7 +50,7 @@ class ESIndexFuzzyMatchedCitations(BaseOperator):
 
     def execute(self, context):
         es = reach.elastic.common.connect(
-            self.es_host, self.es_port)
+            self.es_hosts, self.es_port)
         s3 = WellcomeS3Hook()
 
         # TODO: implement skipping mechanism