--- a/reach/airflow/tasks/es_index_fuzzy_matched.py
+++ b/reach/airflow/tasks/es_index_fuzzy_matched.py
@@ -50,7 +50,7 @@ class ESIndexFuzzyMatchedCitations(BaseOperator):
 
     def execute(self, context):
         es = reach.elastic.common.connect(
-            self.es_hosts, self.es_port)
+            self.es_hosts)
         s3 = WellcomeS3Hook()
 
         # TODO: implement skipping mechanism