--- a/reach/airflow/tasks/exact_match_refs_operator.py
+++ b/reach/airflow/tasks/exact_match_refs_operator.py
@@ -30,9 +30,9 @@ def yield_publications(s3, publications_path):
                 yield json.loads(line)
 
 class ElasticsearchExactMatcher:
-    def __init__(self, es, es_index, title_length_threshold):
+    def __init__(self, es, es_full_text_index, title_length_threshold):
         self.es = es
-        self.es_index = es_index
+        self.es_full_text_index = es_full_text_index
         self.title_length_threshold = title_length_threshold
 
     def match(self, publication):
@@ -52,7 +52,7 @@ class ElasticsearchExactMatcher:
 
         }
         res = self.es.search(
-            index=self.es_index,
+            index=self.es_full_text_index,
             body=body,
             size=1000 # if there are more than 1000, need to paginate
         )
@@ -78,12 +78,12 @@ class ExactMatchRefsOperator(BaseOperator):
     """
 
     @apply_defaults
-    def __init__(self, es_host, publications_path, exact_matched_references_path, es_index,
+    def __init__(self, es_host, publications_path, exact_matched_references_path, es_full_text_index,
                  title_length_threshold, aws_conn_id='aws_default', *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.publications_path = publications_path
         self.exact_matched_references_path = exact_matched_references_path
-        self.es_index = es_index
+        self.es_full_text_index = es_full_text_index
         self.title_length_threshold = title_length_threshold
         self.es_host = es_host
         self.es = Elasticsearch([self.es_host])
@@ -105,7 +105,7 @@ class ExactMatchRefsOperator(BaseOperator):
 
         exact_matcher = ElasticsearchExactMatcher(
             self.es,
-            self.es_index,
+            self.es_full_text_index,
             self.title_length_threshold
         )