Fix #202 

# Description

This PR adapts the code to look for exact matches in all the policy text, rather than just in the references sections.

- Rename exact matcher variables to not have anything to do with sections
- Adds an optional argument in `get_scraping_results` in file_manager.py which allows you to input which columns to include in the scraper output dictionary. This will mean we can include the text column needed to exact match on all the text. I did this rather than just updating the global variable `SCRAPING_COLUMNS` with 'text' since for the fuzzy match it is unneeded to store the large text data in memory.
- I create a new function `transform_scraper_text_file` in `refparse.py` to specifically transform the scraped full text for the exact match. This function is v simliar to `transform_scraper_file`, but I chose to create another function rather than using an if statement in the same one - perhaps this would be better though?
- Added the exact matcher task to the policy dag (editing `exact_match_refs_operator.py` and `policy.py`).
- Adds 8 new exact matcher tests
- Changes to `test_pdf_objects` to allow for these tests to be run (or skipped) on a mac (this should be fixed properly at some point, see issue #273 )

## Type of change

Please delete options that are not relevant.

- [x] :bug: Bug fix (Add `Fix #(issue)` to your PR)

## Running time increases a lot!

A ran locally for the 70 documents in the latest MSF scrape, using:
```
python -m policytool.refparse.refparse \
    --scraper-file "s3://datalabs-dev/reach-airflow/output/policy/parsed-pdfs/msf/parsed-pdfs-msf.json.gz" \
    --references-file "s3://datalabs-data/wellcome_publications/uber_api_publications.csv" \
    --model-file "s3://datalabs-data/reference_parser_models/reference_parser_pipeline.pkl" \
    --output-url "file://./tmp/parser-output/output_folder_name"
```
this took **504 seconds** and finds 10 doc-publication matches.

Without these changes this took **8 seconds** and found 2 doc-publication matches (reassuring also found in the 10 full text search matches).

- I tried flashtext to make it quicker, but it didn't work on looking for whole sentences in large amounts of text - it works for 1 or 2 words in text.
- I tried Whoosh to make this quicker, but it took 1774 seconds and found 7538 matches - obviously there is probably a bug somewhere, but I didn't investigate any further.
- I tried Spacy's PhraseMatcher (`pip install spacy` and `python -m spacy download en_core_web_sm`) but aborted trying this out, since in airflow we dont need it, so it doesnt matter if refparse ran locally is slow

I've created an issue #257, if someone wants to pick it up this performance issue in the future.

# How Has This Been Tested?

```
>>> make docker-test
74 passed, 4 warnings in 10.56s
>>> make test
 72 passed, 2 skipped, 5 warnings in 7.34s
>>> cd reach/refparse/
>>> python -m unittest
Ran 35 tests in 0.112s

OK
```
I ran:
```
source build/virtualenv/bin/activate
eval $(./export_env.py)
docker-compose up -d
./docker_exec.sh airflow test policy-test ExactMatchRefs.msf 2019-11-8
```
it took about 3 hours to run and processed 5377194 publications and found 8132 matches. 
I noticed my results have duplicates in, fixing this might help speed things up if we delete duplicate policy docs before the exact matcher runs. I added some info about this duplication to issue #183