@SamDepardieu 2019-10-07
> this took 504 seconds and finds 10 doc-publication matches.

> Without these changes this took 8 seconds and found 2 doc-publication matches (reassuring also found in the 10 full text search matches).

I don't see anything obvious, but this is slightly worrying in my opinion. 504 seconds for 70 documents is a lot, especially considering that some of our organisations (e.g. parliament) have around 60k documents.

We may want to profile this code to find out what's taking so long and to see if we can come with a solution for it to run faster before merging this PR.

I'd be happy to help you with that if needed, though