@ivyleavedtoadflax 2019-10-07
There is a package I used in a previous role which is significantly faster than regex, and could be used in `ExactMatcher`. I'll see if I can dig it out!