@ivyleavedtoadflax 2019-10-07
Found it We had good results with this. 

https://www.analyticsvidhya.com/blog/2017/11/flashtext-a-library-faster-than-regular-expressions/

https://github.com/vi3k6i5/flashtext