@lizgzil 2019-10-09
@SamDepardieu or @hblanks please can you check my commit changing the airflow stuff if you have time? Tests passed, but I'm not confident the arguments of `exact_match_refs_operator.ExactMatchRefsOperator` are correct.