@lizgzil 2019-10-09
also (@ivyleavedtoadflax @SamDepardieu ) I've given up on making the local running of refparse quicker in this PR since it won't effect the product performance anyway. I'll create an issue, if someone wants to pick it up in the future.