@nsorros 2019-10-10
> @SamDepardieu or @hblanks please can you check my commit changing the airflow stuff if you have time? Tests passed, but I'm not confident the arguments of `exact_match_refs_operator.ExactMatchRefsOperator` are correct.

Did you run the DAG? Did the exact matcher found matches?

Test pass as I think there are no tests testing the exact matcher.