@SamDepardieu 2019-10-10
> > @SamDepardieu or @hblanks please can you check my commit changing the airflow stuff if you have time? Tests passed, but I'm not confident the arguments of `exact_match_refs_operator.ExactMatchRefsOperator` are correct.
> 
> Did you run the DAG? Did the exact matcher found matches?
> 
> Test pass as I think there are no tests testing the exact matcher.

The best way to test the dag is to run the `policy-test` DAG. If you're able to `docker-compose up -d` Reach, you can access Airflow's interface on `localhost:8080` and then just click the little "play" button on the right of the `policy-test` dag