@hblanks 2019-10-31
@lizgzil  - yes, the place for debugging dags is unfortunately in the airflow web (and maybe scheduler logs), using `docker-compose logs airflow-web`, sometimes with options. It's yucky.

Your branch also could stand to be rebased on top of the reach rename. Would you like it if I did this and after we debugged the DAG together?