@lizgzil 2019-10-31
> Your branch also could stand to be rebased on top of the reach rename. Would you like it if I did this and after we debugged the DAG together?

@hblanks Yes please, thatd be great!