@lizgzil 2019-11-08
Thanks to @SamDepardieu and @hblanks I've fixed my issues with Airflow.

I ran `./docker_exec.sh airflow test policy-test ExactMatchRefs.msf 2019-11-8` and got 8132 matches in 3 hours. As noted in the description this takes longer than really necessary since there are duplicates in the ES EPMC Metadata index. I'm currently running the whole policy-test dag, but I anticipate it'll take a long time! (note msf is the smallest data and it still took 3hours).

It'd be great to get this merged so we could see if the policy-test dag succeeds not using my computer, so a review of the code / suggestions on how to speed up performance / help fixing issue #183 would be very much appreciated :)