@hblanks 2019-11-12
@lizgzil  - thanks for keeping this going. A couple thoughts:

1. Because we use the test DAG to quickly verify the pipeline, the test DAG has to complete within a fairly small amount of time.
1. It doesn't sound like we can get the test DAG running quickly right away with the exact matcher in place.
1. But, we do want to have the exact matcher available in the codebase, and maybe even able to run in a deployment environment.

So, my two cents would be that we create a separate dag, `policy-test-exact-match`, which we can use to try speeding up things as best we can. The first place is probably batching search queries to ElasticSearch, if that can be done -- or else running them in a thread pool.