@hblanks 2019-11-14
OK! This branch is rebased to a new branch, `fix-exact-matcher-rebase`. After adding a small commit, the exact matcher test DAG Liz added runs in short order. So, I suggest we:

1. Close this PR.
1. Create a PR for the new branch.
1. Maybe update `dag.py` so that it the main DAG doesn't run the exact matcher, but so that the test DAG does?