@lizgzil 2019-11-14
Closing this and opening a rebased version https://github.com/wellcometrust/reach/pull/285